+++ gemelli/tests/test_preprocessing.py
@@ -7,6 +7,7 @@ from skbio.stats.composition import closure, clr
 
 
 class Testpreprocessing(unittest.TestCase):
+
     def setUp(self):
 
         self.cdata1 = np.array([[2, 2, 6],
