+++ gemelli/tests/test_tenals.py
@@ -1,7 +1,7 @@
 import unittest
 import numpy as np
 from scipy.linalg import qr
-from gemelli.tenals import tenals
+from gemelli.factorization import tenals
 
 
 class Testtenals(unittest.TestCase):
